requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-openai==0.0.8
pandas==2.2.1
python-dotenv==1.0.1
